Bashir Update

+++ mftplus/model/entity/driving_licences.py
@@ -6,11 +6,11 @@ from mftplus.model.entity.base import Base
 class DrivingLicence(Base):
     __tablename__ = "driving_licence_tbl"
     _id = Column("id", Integer, primary_key=True, autoincrement=True)
-    _serial = Column("serial", Integer(30), unique=True)
+    _serial = Column("serial", Integer, unique=True)
     _licence_type = Column("license_type", String(30))
-    _start_date = Column("start_data_time", Integer(20))
-    _expire_date = Column("expire_data_time", Integer(20))
-    _status = Column("status", Integer(20))
+    _start_date = Column("start_data_time", Integer)
+    _expire_date = Column("expire_data_time", Integer)
+    _status = Column("status", Integer)
     _deleted = Column("deleted", Boolean, default=False)
 
     def __init__(self, id, serial, license_type, start_data_time, expire_data_time, status, deleted):
